+++ test-utils/test-utils-core/src/test/java/uk/gov/justice/services/test/utils/core/matchers/EmptyStreamMatcherTest.java
@@ -1,0 +1,22 @@
+package uk.gov.justice.services.test.utils.core.matchers;
+
+import static org.junit.Assert.assertThat;
+import static uk.gov.justice.services.test.utils.core.matchers.EmptyStreamMatcher.isEmptyStream;
+
+import java.util.stream.Stream;
+
+import org.junit.Test;
+
+public class EmptyStreamMatcherTest {
+
+    @Test
+    public void shouldMatchAnEmptyStream() throws Exception {
+        assertThat(Stream.empty(), isEmptyStream());
+    }
+
+    @Test(expected = AssertionError.class)
+    public void shouldNotMatchANonEmptyStream() throws Exception {
+        assertThat(Stream.of("test1", "test2"), isEmptyStream());
+    }
+
+}