+++ CHANGELOG.md
@@ -8,6 +8,7 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 ### Added
 - Support for RestClient to use proxy details from environment properties
 - Matchers for matching handler and handler class methods, annotations and pass through process type (Sender or Requester)
+- Empty stream matching to EventStreamMatcher
 
 ## [0.32.0] - 2016-10-24